+++ crates/libwebrtc-sys/src/cpp/local_audio_source.cc
@@ -63,11 +63,13 @@ void LocalAudioSource::OnData(const void* audio_data,
       webrtc::AudioFrame::SpeechType::kNormalSpeech,
       webrtc::AudioFrame::VADActivity::kVadUnknown, number_of_channels);
 
-  int ap_result = ProcessAudioFrame(audio_processing_.get(), &audio_frame_);
+  if (audio_processing_.get() != nullptr) {
+    int ap_result = ProcessAudioFrame(audio_processing_.get(), &audio_frame_);
 
-  if (ap_result != webrtc::AudioProcessing::kNoError) {
-    RTC_LOG(LS_ERROR) << "`LocalAudioSource`: `AudioProcessing` error: "
-                      << ap_result;
+    if (ap_result != webrtc::AudioProcessing::kNoError) {
+      RTC_LOG(LS_ERROR) << "`LocalAudioSource`: `AudioProcessing` error: "
+                        << ap_result;
+    }
   }
 
   for (auto* sink : sinks_) {