+++ example/integration_test/webrtc_test.dart
@@ -1356,6 +1356,35 @@ void main() {
     }
   });
 
+  testWidgets('Display audio', (WidgetTester tester) async {
+    // Display audio tracks are implemented only on Windows.
+    if (!Platform.isWindows) {
+      return;
+    }
+
+    var caps = DisplayConstraints();
+    caps.video.mandatory = DeviceVideoConstraints();
+    caps.video.mandatory!.width = 640;
+    caps.video.mandatory!.height = 480;
+    caps.audio.mandatory = AudioConstraints();
+
+    var tracks = await getDisplayMedia(caps);
+
+    expect(tracks.length, 2);
+    expect(tracks.where((track) => track.kind() == MediaKind.video).length, 1);
+    expect(tracks.where((track) => track.kind() == MediaKind.audio).length, 1);
+
+    var audioTrack = tracks.firstWhere(
+      (track) => track.kind() == MediaKind.audio,
+    );
+
+    expect(audioTrack.deviceId(), "system_audio_capture");
+
+    for (var track in tracks) {
+      track.dispose();
+    }
+  });
+
   testWidgets('on_track when peer has transceiver.', (
     WidgetTester tester,
   ) async {